feat(quiz): QuizPage implement styles, fix QuizCard style in accordance to design
