+++ src/components/quiz/QuizCard/QuizCard.module.scss
@@ -6,31 +6,47 @@
   align-items: center;
   justify-content: center;
   width: 100%;
-  max-width: 600px; // make sure the card doesn't get too wide on desktop
+  max-width: 900px; // make sure the card doesn't get too wide on desktop
   margin: 0 auto;
   text-align: center;
 
-  animation: fadeIn 0.5s ease-in-out;
+  animation: fadeIn 0.6s cubic-bezier(0.22, 1, 0.36, 1) forwards;
 
   &__question {
     font-family: v.$font-family-base;
-    font-size: 1.5rem;
-    font-weight: v.$font-weight-bold;
+    font-size: 2rem;
     color: v.$color-white;
-    margin-bottom: v.$spacing-lg;
+    margin-bottom: 7rem;
     line-height: 1.2;
+
+    @media (min-width: 768px) {
+      font-size: 3rem;
+      margin-bottom: v.$spacing-xxl;
+    }
   }
 
   &__answers {
     display: flex;
     flex-direction: row;
     justify-content: center;
-    gap: v.$spacing-md;
+    align-items: center;
+    gap: v.$spacing-xl;
     width: 100%;
+    @media (min-width: 768px) {
+      gap: v.$spacing-xxl;
+    }
+  }
+
+  &__button {
+    min-width: 150px;
+
+    @media (min-width: 768px) {
+      min-width: 200px;
+    }
   }
 
   @keyframes fadeIn {
-    from { opacity: 0; transform: translateY(10px); }
+    from { opacity: 0; transform: translateY(30px); }
     to { opacity: 1; transform: translateY(0); }
   }
