+++ src/components/ui/Button/Button.module.scss
@@ -44,6 +44,7 @@
     font-family: v.$font-family-secondary;
     background-color: transparent;
     border-color: v.$color-white;
+    border-style: solid;
     color: v.$color-white;
 
     &:hover {
