+++ src/pages/LandingPage/LandingPage.module.scss
@@ -25,7 +25,7 @@
   position: relative;
   z-index: 10;
   // Account for fixed overlay header + additional spacing
-  padding: calc(v.$header-height + 10vh) 6% 0 6%;
+  padding: calc(#{v.$header-height} + 10vh) 6% 0 6%;
   max-width: 600px; // limit width on desktop
 }
