+++ src/pages/QuizPage/QuizPage.module.scss
@@ -1,0 +1,24 @@
+@use '../../styles/variables' as v;
+
+.page {
+  display: flex;
+  flex-direction: column;
+  color: v.$color-white;
+  padding-top: 10vh;
+  text-align: center;
+}
+
+.content {
+  padding-top: 20vh;
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.subtitle {
+  font-family: v.$font-family-secondary;
+  font-size: 0.75rem;
+  text-transform: uppercase;
+  letter-spacing: 0.6em;
+}
