+++ src/styles/_variables.scss
@@ -14,6 +14,8 @@ $font-weight-medium: 500;
 $spacing-sm: 0.5rem;
 $spacing-md: 1rem;
 $spacing-lg: 2rem;
+$spacing-xl: 3rem;
+$spacing-xxl: 4rem;
 
 /* --- Layout --- */
 $header-height: 72px; // Fixed header height (40px logo + 2rem padding)
